backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
